starter refactoring
