+++ models/contact.js
@@ -1,6 +1,5 @@
 const { Schema, model } = require('mongoose');
 const { handleMongooseError } = require('../helpers');
-const Joi = require('joi');
 
 const emailRegex = /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,3})+$/;
 const phoneRegex =
@@ -37,25 +36,8 @@ const contactSchema = new Schema(
 
 contactSchema.post('save', handleMongooseError);
 
-const addSchema = Joi.object({
-    name: Joi.string().required(),
-    email: Joi.string().email().required(),
-    phone: Joi.string().required(),
-    favorite: Joi.boolean(),
-});
-
-const updateFavoriteSchema = Joi.object({
-    favorite: Joi.boolean().required(),
-});
-
-const schemas = {
-    addSchema,
-    updateFavoriteSchema,
-};
-
 const Contact = model('contact', contactSchema);
 
 module.exports = {
     Contact,
-    schemas,
 };
