+++ routes/api/contacts.js
@@ -9,7 +9,7 @@ const {
     updateFavorite,
 } = require('../../controllers/contacts');
 const { validateBody, isValidId } = require('../../middlewares');
-const { schemas } = require('../../models/contact');
+const { schemas } = require('../../schemas');
 
 const router = express.Router();
