+++ schemas/index.js
@@ -1,0 +1,5 @@
+const schemas = require('./contacts');
+
+module.exports = {
+    schemas,
+};
